fix(viewer): bake the gradient/colormap LUT at f32 (complete the banding fix)
Output dithering alone left residual banding in dark gradients because the
gradient LUT itself was 8-bit: `Gradient::bake` quantized the ramp *values* to
`Vec<u8>` and `make_lut_texture` stored them as `Rgba8Unorm`, so dark ramps —
where sRGB is steepest — carried coarse LUT-side steps the output dither couldn't
recover.

Bake to `Vec<f32>` and store the LUT as `Rgba32Float` (the layout already binds
it filterable; FLOAT32_FILTERABLE is enabled, and the .cube LUT already uses this
format). The ramp is now smooth at the source; combined with the output dither
the dark background gradient renders clean. LUT is 256×1, so ~4 KB — negligible.
Applies to both the background gradient and the diff colormap.

Verified on Metal: default 0.05→0.3 background gradient renders smooth, no
validation errors from the f32 LUT upload.

+++ src/viewer.rs
@@ -530,10 +530,10 @@ pub struct ExrViewer {
     /// User-saved named gradients (the preset library shared with the gradient
     /// editor). Round-tripped through `ExrApp` for persistence.
     pub custom_gradients: Vec<(String, Gradient)>,
-    /// Baked 256-entry colormap LUT bytes + the colormap they were baked from, so
-    /// the GPU texture is re-uploaded only when the active gradient changes.
-    /// Transient (rebuilt on demand).
-    colormap_lut: Vec<u8>,
+    /// Baked 256-entry colormap LUT (f32 texels) + the colormap they were baked
+    /// from, so the GPU texture is re-uploaded only when the active gradient
+    /// changes. Transient (rebuilt on demand).
+    colormap_lut: Vec<f32>,
     colormap_sig: Option<Colormap>,
     /// Transient gradient-editor window state. Shared by the diff colormap editor
     /// and the background gradient editor; `gradient_editor_target` says which.
@@ -552,9 +552,9 @@ pub struct ExrViewer {
     /// name. Transient.
     pub show_background_window: bool,
     new_bg_preset_name: String,
-    /// Baked background-gradient LUT bytes + the ramp they were baked from, so the
-    /// GPU texture is re-uploaded only when the gradient ramp changes.
-    bg_gradient_lut: Vec<u8>,
+    /// Baked background-gradient LUT (f32 texels) + the ramp they were baked from,
+    /// so the GPU texture is re-uploaded only when the gradient ramp changes.
+    bg_gradient_lut: Vec<f32>,
     bg_gradient_sig: Option<Gradient>,
     pub blink_state: bool,
     pub blink_interval: f32,
